Fix incorrect epoch start time, allow configuration.
A mistake was made in the epoch start time. It was defined in seconds,
instead of milliseconds. As a result, the time component of previously
generated IDs was calculated incorrectly and the time period able to be
represented by the IDs was only possible via wrapping. As a result, the
epoch needs to be changed to be in the expected format so that the
gurantees hold.