Make sure loading Lua from resources works.
Unfortunately, the switch away from `IOUtils` has actually broken the
reading of the Lua script from resources when packed into an external
project.

Thankfully, Java 8 has ways now to read the resource as a stream and not
as a file without writing a billion lines of code, so we'll switch to
that.